refactor: use macro to further simplify distributor startup
