+++ src/formatters/visa_settlement.rs
@@ -1,64 +1,10 @@
 use crate::{formatters::to_pounds, models::Transaction};
-use chrono::Utc;
 use color_eyre::{eyre::eyre, Result};
 use serde_json::json;
 
-use super::Formatter;
+pub struct Formatter;
 
-pub struct AuthFormatter;
-
-impl Formatter for AuthFormatter {
-    fn format(transactions: Vec<Transaction>) -> Result<String> {
-        let transaction = transactions
-            .into_iter()
-            .next()
-            .ok_or_else(|| eyre!("Expected at least one transaction."))?;
-
-        let amount = to_pounds(transaction.amount);
-        let date = transaction.transaction_date.to_rfc3339();
-        let auth = json!({
-            "CardId": transaction.transaction_id[0..9],
-            "ExternalUserId": transaction.token,
-            "MessageElementsCollection": [
-                {"Key": "Transaction.BillingAmount", "Value": amount},
-                {"Key": "Transaction.TimeStampYYMMDD", "Value": date},
-                {"Key": "Transaction.MerchantCardAcceptorId", "Value": primary_identifier(&transaction)},
-                {"Key": "Transaction.MerchantAcquirerBin", "Value": "3423432"},
-                {"Key": "Transaction.TransactionAmount", "Value": amount},
-                {"Key": "Transaction.VipTransactionId", "Value": transaction.transaction_id},
-                {"Key": "Transaction.VisaMerchantName", "Value": transaction.merchant_name},
-                {"Key": "Transaction.VisaMerchantId", "Value": psimi_identifier(&transaction)},
-                {"Key": "Transaction.VisaStoreName", "Value": "Bink Shop"},
-                {"Key": "Transaction.VisaStoreId", "Value": secondary_identifier(&transaction)},
-                {"Key": "Transaction.SettlementDate", "Value": ""},
-                {"Key": "Transaction.SettlementAmount", "Value": 0},
-                {"Key": "Transaction.SettlementCurrencyCodeNumeric", "Value": 0},
-                {"Key": "Transaction.SettlementBillingAmount", "Value": 0},
-                {"Key": "Transaction.SettlementBillingCurrency", "Value": ""},
-                {"Key": "Transaction.SettlementUSDAmount", "Value": 0},
-                {"Key": "Transaction.CurrencyCodeNumeric", "Value": "840"},
-                {"Key": "Transaction.BillingCurrencyCode", "Value": "840"},
-                {"Key": "Transaction.USDAmount", "Value": amount},
-                {"Key": "Transaction.MerchantLocalPurchaseDate ", "Value": "2019-12-19"},
-                {"Key": "Transaction.MerchantGroup.0.Name", "Value": "TEST_MG"},
-                {"Key": "Transaction.MerchantGroup.0.ExternalId", "Value": "MYSTORE"},
-                {"Key": "Transaction.MerchantDateTimeGMT ", "Value": date},
-                {"Key": "Transaction.AuthCode", "Value": transaction.auth_code},
-                {"Key": "Transaction.PanLastFour", "Value": transaction.last_four},
-            ],
-            "MessageId": "12345678",
-            "MessageName": "AuthMessageTest",
-            "UserDefinedFieldsCollection": [{"Key": "TransactionType", "Value": "AUTH"}],
-            "UserProfileId": "f292f99d-babf-528a-8d8a-19fa5f14f4"
-        });
-
-        Ok(auth.to_string())
-    }
-}
-
-pub struct SettlementFormatter;
-
-impl Formatter for SettlementFormatter {
+impl super::Formatter for Formatter {
     fn format(transactions: Vec<Transaction>) -> Result<String> {
         let transaction = transactions
             .into_iter()
@@ -82,7 +28,7 @@ impl Formatter for SettlementFormatter {
                     {"Key": "Transaction.VisaMerchantId", "Value": psimi_identifier(&transaction)},
                     {"Key": "Transaction.VisaStoreName", "Value": "Bink Shop"},
                     {"Key": "Transaction.VisaStoreId", "Value": secondary_identifier(&transaction)},
-                    {"Key": "Transaction.SettlementDate", "Value": Utc::now()},
+                    {"Key": "Transaction.SettlementDate", "Value": date},
                     {"Key": "Transaction.SettlementAmount", "Value": amount},
                     {"Key": "Transaction.SettlementCurrencyCodeNumeric", "Value": 826},
                     {"Key": "Transaction.SettlementBillingAmount", "Value": amount},
@@ -134,6 +80,8 @@ mod tests {
     use chrono::Utc;
     use pretty_assertions::assert_eq;
 
+    use crate::formatters::Formatter as _;
+
     use super::*;
 
     #[test]
@@ -143,7 +91,7 @@ mod tests {
     }
 
     #[test]
-    fn visa_auth_valid() -> Result<()> {
+    fn visa_settlement_valid() -> Result<()> {
         let dt = Utc::now();
         let test_transaction = Transaction {
             amount: 245,
@@ -159,7 +107,7 @@ mod tests {
             last_four: "7890".to_owned(),
         };
 
-        let json_result = AuthFormatter::format(vec![test_transaction]);
+        let json_result = Formatter::format(vec![test_transaction]);
         assert_eq!(
             serde_json::from_str::<serde_json::Value>(&json_result?)?,
             json!({
@@ -176,25 +124,25 @@ mod tests {
                     {"Key": "Transaction.VisaMerchantId", "Value": "PSIMI3333"},
                     {"Key": "Transaction.VisaStoreName", "Value": "Bink Shop"},
                     {"Key": "Transaction.VisaStoreId", "Value": "SEC222222"},
-                    {"Key": "Transaction.SettlementDate", "Value": ""},
-                    {"Key": "Transaction.SettlementAmount", "Value": 0},
-                    {"Key": "Transaction.SettlementCurrencyCodeNumeric", "Value": 0},
-                    {"Key": "Transaction.SettlementBillingAmount", "Value": 0},
-                    {"Key": "Transaction.SettlementBillingCurrency", "Value": ""},
-                    {"Key": "Transaction.SettlementUSDAmount", "Value": 0},
+                    {"Key": "Transaction.SettlementDate", "Value": dt.to_rfc3339()},
+                    {"Key": "Transaction.SettlementAmount", "Value": "2.45"},
+                    {"Key": "Transaction.SettlementCurrencyCodeNumeric", "Value": 826},
+                    {"Key": "Transaction.SettlementBillingAmount", "Value": "2.45"},
+                    {"Key": "Transaction.SettlementBillingCurrency", "Value": "GBP"},
+                    {"Key": "Transaction.SettlementUSDAmount", "Value": "2.45"},
                     {"Key": "Transaction.CurrencyCodeNumeric", "Value": "840"},
                     {"Key": "Transaction.BillingCurrencyCode", "Value": "840"},
                     {"Key": "Transaction.USDAmount", "Value": "2.45"},
-                    {"Key": "Transaction.MerchantLocalPurchaseDate ", "Value": "2019-12-19"},
+                    {"Key": "Transaction.MerchantLocalPurchaseDate", "Value": "2019-12-19"},
                     {"Key": "Transaction.MerchantGroup.0.Name", "Value": "TEST_MG"},
                     {"Key": "Transaction.MerchantGroup.0.ExternalId", "Value": "MYSTORE"},
-                    {"Key": "Transaction.MerchantDateTimeGMT ", "Value": dt.to_rfc3339()},
+                    {"Key": "Transaction.MerchantDateTimeGMT", "Value": dt.to_rfc3339()},
                     {"Key": "Transaction.AuthCode", "Value": "123456"},
                     {"Key": "Transaction.PanLastFour", "Value": "7890"},
                     ],
                     "MessageId": "12345678",
-                    "MessageName": "AuthMessageTest",
-                    "UserDefinedFieldsCollection": [{"Key": "TransactionType", "Value": "AUTH"}],
+                    "MessageName": "SettlementMessageTest",
+                    "UserDefinedFieldsCollection": [{"Key": "TransactionType", "Value": "SETTLE"}],
                     "UserProfileId": "f292f99d-babf-528a-8d8a-19fa5f14f4"
             })
         );
